fix: consolidate packaged TTS metrics in service layer
Remove duplicate packaged-factory metric emission from the CLI wrapper so
one invocation records attempts and terminal outcomes only in the service
wait loop. Split oversized backend files to satisfy the backend-size gate.
