+++ pkg/service/factory_runtime_state.go
@@ -955,49 +955,3 @@ func runtimeModeOrDefault(mode interfaces.RuntimeMode) interfaces.RuntimeMode {
 	}
 	return mode
 }
-
-func (fs *FactoryService) waitForServiceModeStartupWorkReadability(ctx context.Context, serviceMode bool) error {
-	policy := fs.coordinatorPolicy()
-	if !serviceMode || policy.workFile == "" || policy.apiServerReady == nil || policy.port <= 0 || policy.apiServerStarter == nil {
-		return nil
-	}
-	apiServerExit := fs.apiServerExit
-	select {
-	case <-policy.apiServerReady:
-	case err := <-apiServerExit:
-		return startupReadinessError(err)
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-
-	timer := time.NewTimer(serviceModeStartupWorkReadabilityDelay)
-	defer timer.Stop()
-
-	select {
-	case <-timer.C:
-		return nil
-	case err := <-apiServerExit:
-		return startupReadinessError(err)
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-}
-
-func (fs *FactoryService) failServiceModeStartup(currentRuntime *liveRuntimeHandle, startupErr error) error {
-	fs.clearRunState()
-	fs.unregisterLiveSession(defaultFactorySessionID)
-	if currentRuntime == nil {
-		return startupErr
-	}
-	if stopErr := fs.stopLiveRuntime(currentRuntime); stopErr != nil && !errors.Is(stopErr, context.Canceled) {
-		return errors.Join(startupErr, stopErr)
-	}
-	return startupErr
-}
-
-func startupReadinessError(err error) error {
-	if err == nil {
-		return fmt.Errorf("wait for service-mode startup work readiness: API server stopped before signaling readiness")
-	}
-	return fmt.Errorf("wait for service-mode startup work readiness: %w", err)
-}
